+++ src/domain/johtoselvitys_new/johtoselvitysCreateDialog/JohtoselvitysCreateDialog.tsx
@@ -1,25 +1,17 @@
 import { Box } from '@chakra-ui/react';
-import {
-  Button,
-  Dialog,
-  TextInput as HDSTextInput,
-  IconCheck,
-  IconCross,
-  IconInfoCircle,
-  Notification,
-} from 'hds-react';
+import { Button, Dialog, IconCheck, IconCross, IconInfoCircle, Notification } from 'hds-react';
 import { useTranslation } from 'react-i18next';
 import { useNavigate } from 'react-router-dom';
 import { FormProvider, useForm } from 'react-hook-form';
 import { useMutation } from 'react-query';
 import { yupResolver } from '@hookform/resolvers/yup';
 import TextInput from '../../../common/components/textInput/TextInput';
-import useUser from '../../auth/useUser';
 import useLinkPath from '../../../common/hooks/useLinkPath';
 import { ROUTES } from '../../../common/types/route';
 import { NewJohtoselvitysData } from '../../application/types/application';
 import { newJohtoselvitysSchema } from '../validationSchema';
 import { createJohtoselvitys } from '../../application/utils';
+import OwnInformationFields from '../../forms/components/OwnInformationFields';
 
 type Props = {
   isOpen: boolean;
@@ -28,7 +20,6 @@ type Props = {
 
 function JohtoselvitysCreateDialog({ isOpen, onClose }: Readonly<Props>) {
   const { t } = useTranslation();
-  const user = useUser();
   const navigate = useNavigate();
   const getEditJohtoselvitysPath = useLinkPath(ROUTES.EDIT_JOHTOSELVITYSHAKEMUS);
   const formContext = useForm<NewJohtoselvitysData>({
@@ -73,34 +64,8 @@ function JohtoselvitysCreateDialog({ isOpen, onClose }: Readonly<Props>) {
       <FormProvider {...formContext}>
         <form onSubmit={handleSubmit(submitForm)}>
           <Dialog.Content>
-            <Box marginBottom="var(--spacing-s)">
-              <h3 className="heading-s">{t('form:labels:omatTiedot')}</h3>
-            </Box>
-            <Box marginBottom="var(--spacing-m)">
-              <HDSTextInput
-                id="user-name"
-                label={t('form:yhteystiedot:labels:nimi')}
-                value={`${user.data?.profile?.name}`}
-                helperText={t('form:labels:fromHelsinkiProfile')}
-                readOnly
-              />
-            </Box>
-            <Box marginBottom="var(--spacing-s)" maxWidth={328}>
-              <TextInput
-                name="perustaja.sahkoposti"
-                label={t('hankeForm:labels:email')}
-                required
-                defaultValue={user.data?.profile.email}
-              />
-            </Box>
-            <Box marginBottom="var(--spacing-m)" maxWidth={328}>
-              <TextInput
-                name="perustaja.puhelinnumero"
-                label={t('hankeForm:labels:puhelinnumero')}
-                required
-              />
-            </Box>
-            <Box marginBottom="var(--spacing-s)">
+            <OwnInformationFields />
+            <Box marginTop="var(--spacing-m)" marginBottom="var(--spacing-s)">
               <h3 className="heading-s">{t('hakemus:labels:applicationInfo')}</h3>
             </Box>
             <Box marginBottom="var(--spacing-2-xs)">
